Sof Act 1.5

+++ backend/RoyServices/RoyServices/src/main/java/com/example/RoyServices/dto/ObjetoConZonaProjection.java
@@ -1,16 +1,13 @@
-
 package com.example.RoyServices.dto;
-
-import java.math.BigDecimal;
 
 public interface ObjetoConZonaProjection {
     Integer getIdObjeto();
     Integer getIdUsArrendador();
     String getNombreObjeto();
-    BigDecimal getPrecio();
+    Double getPrecio();
     String getEstado();
     String getCategoria();
     String getDescripcion();
     String getImagenUrl();
-    String getZona(); // viene de Usuario
+    String getZona();
 }
