+++ backend/RoyServices/RoyServices/src/main/java/com/example/RoyServices/dto/ObjetoDto.java
@@ -2,7 +2,6 @@ package com.example.RoyServices.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -10,7 +9,7 @@ public class ObjetoDto {
     private Integer idObjeto;
     private Integer idUsArrendador;
     private String nombreObjeto;
-    private BigDecimal precio;
+    private Double precio;
     private String estado;
     private String categoria;
     private String descripcion;
